+++ src/components/Modal/Component.ts
@@ -1,39 +1,29 @@
 import Button from "@elements/Button";
-import { $divApp } from "@constants/div.app";
-import Typography from "@elements/Typography";
+import { $app } from "@constants/div.app";
 
-import {
-  type ModalImplements,
-  type ModalConstructor
-} from "./type";
+import { type ModalImplements } from "./type";
 
 import style from './style.module.scss';
 
 class Modal implements ModalImplements {
   protected $modalWrapper: HTMLElement | null = null;
+  protected $modalTitle: HTMLElement | null = null;
   protected $modalBody: HTMLElement | null = null;
   protected $modalFooter: HTMLElement | null = null;
-  protected title: Typography | null = null;
 
-  constructor({
-    title = '',
-  }: ModalConstructor) {
-    if (title) {
-      this.title = new Typography({
-        text: title,
-        type: 'h6',
-        textColor: 'black',
-        textWeight: 500
-      });
-    }
+  constructor() {
   }
 
   public get modal() {
     return this.$modalWrapper;
   }
 
+  public get modalTitle() {
+    return this.$modalTitle;
+  }
+
   public get modalBody() {
-    return this.$modalWrapper;
+    return this.$modalBody;
   }
 
   public get modalFooter() {
@@ -47,7 +37,7 @@ class Modal implements ModalImplements {
     this.$modalWrapper = $modalWrapper;
 
     $modalWrapper.appendChild(this.buildModalOverlay());
-    $divApp?.appendChild(this.$modalWrapper);
+    $app?.appendChild(this.$modalWrapper);
   }
 
   protected buildModalOverlay = () => {
@@ -99,9 +89,6 @@ class Modal implements ModalImplements {
       'align-items-center'
     ].join(' ')
     ;
-    if (this.title?.textElement) {
-    $title.appendChild(this.title?.textElement);
-    }
 
     const $modalCross = new Button({
       startIcon: {
@@ -117,6 +104,8 @@ class Modal implements ModalImplements {
     if ($modalCross.buttonElement) {
     $title.appendChild($modalCross.buttonElement)
     }
+
+    this.$modalTitle = $title
 
     return $title;
   }
@@ -155,17 +144,24 @@ class Modal implements ModalImplements {
     }
   }
 
-  public openModal = ($body: HTMLElement | null, $footer: HTMLElement | null) => {
+  public openModal = ($title: HTMLElement | null, $body: HTMLElement | null, $footer: HTMLElement | null) => {
     this.toggleBodyHide();
 
     this.buildModalWrapper();
+
+    const $btnCross = this.$modalTitle?.querySelector('button')
+
+    if ($title  && $btnCross) {
+      this.$modalTitle?.insertBefore($title, $btnCross);
+      }
+
     if ($body) {
     this.$modalBody?.appendChild($body);
     }
 
     if ($footer) {
-      this.$modalBody?.appendChild($footer);
-      }
+      this.$modalFooter?.appendChild($footer);
+    }
   }
 
   protected destroy = () => {
