+++ src/modules/AddProductForm/index.ts
@@ -1,3 +1,1 @@
 export { default } from './Component'
-
-export { type AddProductFormImplements} from './type'
