+++ src/modules/ProductForm/index.ts
@@ -1,0 +1,3 @@
+export { default } from './Component'
+
+export { type ProductFormImplements} from './type'
