+++ src/modules/ProductForm/type.ts
@@ -1,0 +1,3 @@
+export interface ProductFormImplements {
+  formProduct: HTMLElement | null
+}
